refactor(rvr): update rvr to have 64bit PC (#2923)
update rvr to have 64bit PC. This PR consist of multiple parts:
1) Updating Intermediate Representations to have 64bit PC instead of
32bit
2) Updating Lifter to use 64bit PC
3) Updating Emitter to use 64bit PC for its function arguments, update
PC related constants to 64bit.
4) Adding boundary checks during compile-time for statically computable
PCs (`JAL` and branching instructions)

resolves INT-8319

+++ crates/rvr/rvr-openvm/c/suspender/openvm_suspender_instret_limit.h
@@ -8,7 +8,7 @@
 #include "openvm_state.h"
 
 static __attribute__((always_inline)) inline uint8_t should_suspend(
-    RvState* restrict state, uint32_t pc, uint32_t block_insn_count,
+    RvState* restrict state, uint64_t pc, uint32_t block_insn_count,
     uint8_t suspend_signal) {
   if (unlikely(state->instret > state->target_instret)) {
     state->instret -= block_insn_count;
